requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.5
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
config==0.4.2
cryptography==2.4.2
Django==2.2.1
Flask==1.0.2
httplib2==0.17.0
idna==2.8
image==1.5.28
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
paramiko==2.4.2
Pillow==7.0.0
pyasn1==0.4.5
pycparser==2.19
pymongo==3.10.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyTelegramBotAPI==3.6.7
pytz==2019.1
readline==6.2.4.1
redis==3.2.1
requests==2.7.0
six==1.12.0
sqlparse==0.3.0
tencentcloud-sdk-python==3.0.125
urllib3==1.24.1
uWSGI==2.0.18
virtualenv==16.7.5
Werkzeug==0.14.1
